Reworked build infra
The old build script was lengthy, not easily changeable, difficult to
read. I rewrote the build scripts into one (hpvm_build.sh) with this
in mind.

In the past, everything in llvm_patches gets placed at the same relative path in
llvm EXCEPT include headers. I changed this to be more consistent, not
hard-code paths, and make it easier to add a new file to llvm_patches
in the future (it will be automatically picked up).

I added infrastructure to build in a docker container. I use a docker
script instead of docker container, so the process in the container
can adopt the user is running on the host system. Now the container
can write to the volume as the real user--not as root.